+++ YoloOnnxWinform/YoloDotNet/Modules/12/ObjectDetectionModuleV12.cs
@@ -23,7 +23,7 @@ namespace YoloDotNet.Modules.V12
             _objectDetectionModuleV8 = new ObjectDetectionModuleV8(_yoloCore);
         }
 
-        public List<ObjectDetection> ProcessImage<T>(T image, double confidence, double pixelConfidence, double iou)
+        public List<ObjectDetection> ProcessImage<T>(T image, float confidence, float pixelConfidence, float iou)
             => _objectDetectionModuleV8.ProcessImage(image, confidence, pixelConfidence, iou);
 
         public void Dispose()
