+++ YoloOnnxWinform/YoloDotNet/Modules/Interfaces/ISegmentationModule.cs
@@ -6,6 +6,6 @@ namespace YoloDotNet.Modules.Interfaces
 {
     public interface ISegmentationModule : IModule
     {
-        List<Segmentation> ProcessImage<T>(T image, double confidence, double pixelConfidence, double iou);
+        List<Segmentation> ProcessImage<T>(T image, float confidence, float pixelConfidence, float iou);
     }
 }
